requirements.txt
youtube_transcript_api
streamlit
pathlib
python-dotenv
google-generativeai